updated after PR comments
