+++ cypress/fixtures/products.json
@@ -4,7 +4,19 @@
       "id": 1,
       "title": "pizza",
       "description": "best pizza in town",
-      "price": 99
+      "price": 10
+    },
+    {
+      "id": 2,
+      "title": "hamburger",
+      "description": "everything on it",
+      "price": 6
+    },
+    {
+      "id": 3,
+      "title": "kebab",
+      "description": "kebab in pita bread",
+      "price": 7
     }
   ]
 }
