+++ src/App.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import ViewMenu from "./components/ViewMenu";
+import Menu from "./components/Menu";
 
 const App = () => {
   return (
     <>
       <h1>Turtle - Food fast</h1>
-      <ViewMenu />
+      <Menu />
     </>
   );
 };
