+++ src/modules/menu.js
@@ -1,10 +1,6 @@
 import axios from "axios";
 
 const getMenu = async () => {
-  // let headers = {
-  //   "content-type": "application/json",
-  //   Accept: "application/json",
-  // };
   const response = await axios.get("/api/v1/products");
   return response;
 };
